Updates per review comments - clean up CRD comments, and remove obsolete field - rename the instance package/files
