+++ platform-operator/config/rbac/role.yaml
@@ -1,0 +1,28 @@
+
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  creationTimestamp: null
+  name: manager-role
+rules:
+- apiGroups:
+  - clusters.verrazzano.io
+  resources:
+  - verrazzanomanagedclusters
+  verbs:
+  - create
+  - delete
+  - get
+  - list
+  - patch
+  - update
+  - watch
+- apiGroups:
+  - clusters.verrazzano.io
+  resources:
+  - verrazzanomanagedclusters/status
+  verbs:
+  - get
+  - patch
+  - update
