+++ platform-operator/internal/vzinstance/verrazzano_instance.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package instance
+package vzinstance
 
 import (
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
