+++ platform-operator/internal/vzinstance/verrazzano_instance_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2021, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package instance
+package vzinstance
 
 import (
 	"fmt"
